addsite: run installdbs as the application user to fix logging
Currently, the installdbs script is run as the root user in addsite.
This doesn't appear to be necessary for any specific reason, and results
in the application not having permission to write to the created log
file. Updating to run as the application user fixes logging and doesn't
appear to have any negative side effects.

Hygiene: Also removes an extra trailing slash from the
$dockerLogDirectory value set in config/mediawiki/LocalSettings.php.